Simplified bld script and added stricter file types
