+++ hat/hatrun
@@ -1,4 +1,4 @@
-/*
+/* vim: set ft=java:
  * Copyright (c) 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
@@ -25,67 +25,68 @@
 
 import static bldr.Bldr.*;           // all the helpers are here 
 import static java.nio.file.Files.*; // so we can use isDirectory(path);
+ 
 
 void main(String[] args) {
-   var hatDir = new Root();
-   if (args.length==0){
-      println("""
-              usage:
-                 hatrun [headless] backend package args ...
-                 headless : Optional passes -Dheadless=true to app
-                 backend  : opencl|cuda|spirv|ptx|mock
-                 package  : the examples package (and dirname under hat/examples)
-                 Class name is assumed to be package.Main 
-              """);
-      System.exit(1);
-   } else{
-     var buildDir = assertOrThrow(hatDir.buildDir(), Files::isDirectory, "Build Directory");
-     
-     int argc=0;
-     boolean headless = args[argc].equals("headless");
-     if (headless){
-        argc++;
-     }
-     if (argc<args.length 
-           && hatDir.path().resolve("backends") instanceof Path backendsDir && Files.isDirectory(backendsDir)){
-        var classPath = new ArrayList<>(List.of(buildDir.resolve("hat-1.0.jar")));
-        var backendDir = backendsDir.resolve(args[argc++]);
-        if (argc<args.length && (Files.isDirectory(backendDir) || backendDir.getFileName().toString().equals("java"))){
-           if (Files.isDirectory(backendDir)){
-              classPath.add(assertOrThrow(buildDir.resolve("hat-backend-"+backendDir.getFileName()+"-1.0.jar"), Files::isRegularFile," backend dir"));
-           }else{
-              classPath.add(assertOrThrow(backendsDir.resolve("shared/src/main/resources"), Files::isDirectory, "backend dir"));
-           }
-           if (argc<args.length &&
-                assertOrThrow(hatDir.path().resolve("examples"), Files::isDirectory, "Examples Dir") instanceof Path examplesDir){
-              var exampleDir = assertOrThrow(examplesDir.resolve(args[argc++]), Files::isDirectory, "Example dir");
-              classPath.add(assertOrThrow(buildDir.resolve("hat-example-"+exampleDir.getFileName()+"-1.0.jar"), Files::isRegularFile," example jar "));
-              final int finalArgc = argc; //sigh
-              
-              java($->$
-                 .vmopts(
-                   "--enable-preview",
-                   "--enable-native-access=ALL-UNNAMED",
-                   "--add-exports=java.base/jdk.internal=ALL-UNNAMED",
-                   "-Djava.library.path=build:/usr/local/lib"
-                 )
-                 .when(headless, $$->$$
-                    .vmopts("-Dheadless=true")
-                 )
-                 .class_path(classPath)
-                 .main_class(exampleDir.getFileName() + ".Main")
-                 .args(Arrays.copyOfRange(args, finalArgc, args.length))
-              );
- 
-           }else{
-              println("no example");
-           }
-        }else{
-           println("no backend");
-        }
-     }else{
-        println("junk");
-     }
+ var usage =
+    """
+    usage:
+      java @bldr/args hatrun [headless] backend package args ...
+         [headless] : Optional passes -Dheadless=true to app
+          backend   : opencl|cuda|spirv|ptx|mock
+          package   : the examples package (and dirname under hat/examples)
 
-   }
+      class name is assumed to be package.Main  (i.e. mandel.main) 
+
+      examples:
+         java @bldr/args opencl mandel
+         java @bldr/args headless opencl mandel
+    """;
+
+  var hatDir =  Dir.current();
+  var backends = assertExists(hatDir.dir("backends"));
+  var examples = assertExists(hatDir.dir("examples"));
+  var buildDir = assertExists(BuildDir.of(hatDir.path("build")));
+
+  var javaBuilder = javaBuilder().vmopts(
+                  "--enable-preview",
+                  "--enable-native-access=ALL-UNNAMED",
+                  "--add-exports=java.base/jdk.internal=ALL-UNNAMED"
+          )
+          .library_path(buildDir)
+          .class_path(buildDir.jarFile("hat-1.0.jar"));
+
+  int argn = 0;
+  if (args.length > 0 && args[argn].equals("headless")) {
+      javaBuilder.vmopts("-Dheadless=true");
+      argn++;
+  }
+  if ((argn + 2) > args.length) {
+      print("args[" + args.length + "] = [ ");
+      List.of(args).forEach(a -> print(" " + a));
+      println(" ]");
+      println(usage);
+  } else {
+      var backendName = args[argn++];
+      var exampleName = args[argn++];
+
+      if (backendName.equals("java")) {
+          javaBuilder.class_path(ClassDir.of(backends.path("shared/src/main/resources")));
+      } else if (backends.dir(backendName) instanceof Dir backend && backend.exists()) {
+          javaBuilder.class_path(buildDir.jarFile("hat-backend-" + backendName + "-1.0.jar"));
+      } else {
+          println("No backend " + backendName);
+      }
+
+      if (examples.dir(exampleName) instanceof Dir example && example.exists()) {
+          java(javaBuilder
+                  .verbose()
+                  .class_path(buildDir.jarFile("hat-example-" + exampleName + "-1.0.jar"))
+                  .main_class(exampleName + ".Main")
+                  .args(Arrays.copyOfRange(args, argn, args.length))
+          );
+      } else {
+          println("no example " + exampleName);
+      }
+  }
 }
